requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.2
django-environ==0.9.0
django-filter==22.1
django-password-validators==1.7.0
django-regex==0.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.2.0
PyJWT==2.6.0
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.5
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
